+++ src/utils/observable.js
@@ -1,6 +1,9 @@
 import { Observable } from 'rxjs/Observable';
 import { toPromise } from 'rxjs/operator/toPromise';
 
+/**
+ * @private
+ */
 export class KinveyObservable extends Observable {
   toPromise() {
     return toPromise.call(this);