+++ src/utils/storage.js
@@ -1,15 +1,16 @@
 import localStorage from 'local-storage';
 const userCollectionName = process.env.KINVEY_USER_COLLECTION_NAME || 'kinvey_user';
-const socialIdentityCollectionName = process.env.KINVEY_SOCIAL_IDENTITY_COLLECTION_NAME
-                                                || 'kinvey_socialIdentity';
-const syncKeyCollectionName = process.env.KINVEY_SYNC_KEY_COLLECTION_NAME || 'kinvey_syncKey';
 
-// Get the active user
+/**
+ * @private
+ */
 export function getActiveUser(client) {
   return localStorage.get(`${client.appKey}${userCollectionName}`);
 }
 
-// Set the active user
+/**
+ * @private
+ */
 export function setActiveUser(client, data) {
   if (data) {
     try {
@@ -22,38 +23,24 @@ export function setActiveUser(client, data) {
   return localStorage.remove(`${client.appKey}${userCollectionName}`);
 }
 
-// Get the active social identity
-export function getActiveSocialIdentity(client) {
-  return localStorage.get(`${client.appKey}${socialIdentityCollectionName}`);
+/**
+ * @private
+ */
+export function getIdentitySession(client, identity) {
+  return localStorage.get(`${client.appKey}${identity}`);
 }
 
-// Set the active social identity
-export function setActiveSocialIdentity(client, data) {
-  if (data) {
+/**
+ * @private
+ */
+export function setIdentitySession(client, identity, session) {
+  if (session) {
     try {
-      return localStorage.set(`${client.appKey}${socialIdentityCollectionName}`, data);
+      return localStorage.set(`${client.appKey}${identity}`, session);
     } catch (error) {
       return false;
     }
   }
 
-  return localStorage.remove(`${client.appKey}${socialIdentityCollectionName}`);
-}
-
-// Get the sync key
-export function getSyncKey(client) {
-  return localStorage.get(`${client.appKey}${syncKeyCollectionName}`);
-}
-
-// Set the sync key
-export function setSyncKey(client, key) {
-  if (key) {
-    try {
-      return localStorage.set(`${client.appKey}${syncKeyCollectionName}`, key);
-    } catch (error) {
-      return false;
-    }
-  }
-
-  return localStorage.remove(`${client.appKey}${syncKeyCollectionName}`);
+  return localStorage.remove(`${client.appKey}${identity}`);
 }